pyproject.toml
[tool.poetry]
name = "package-installer"
version = "0.1.0"
description = ""
authors = ["\"Joris de Josselin de Jong\" <joor-josselin@hotmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
typer = "^0.12.5"
gitpython = "^3.1.43"
pyinstaller = "^6.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
build = "build:build_executable"